+++ src/backend/Views/Word/form_new.php
@@ -66,6 +66,12 @@ data-lwt-clear-frame="true">
             </td>
         </tr>
         <tr>
+            <td class="td1 right">Notes:</td>
+            <td class="td1">
+                <textarea name="WoNotes" id="WoNotes" cols="35" rows="3" class="textarea-noreturn checklength checkoutsidebmp" data_maxlength="1000" data_info="Notes"></textarea>
+            </td>
+        </tr>
+        <tr>
             <td class="td1 right">Status:</td>
             <td class="td1">
                 <?php echo SelectOptionsBuilder::forWordStatusRadio(1); ?>